# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.CreateUserSubscriptionRequest:
    properties:
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      notes:
        example: Admin assigned subscription
        type: string
      server_group_ids:
        items:
          type: integer
        type: array
      skip_payment:
        example: true
        type: boolean
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
      subscription_plan_id:
        example: 1
        type: integer
      use_trial:
        example: false
        type: boolean
      user_id:
        example: 1
        type: integer
    required:
    - subscription_plan_id
    - user_id
    type: object
...
  response.BadRequestResponse:
    description: Bad Request response format
    properties:
      code:
        example: 4000
        type: integer
      message:
        example: Invalid request parameters
        type: string
    type: object
...
  response.ForbiddenResponse:
    description: Forbidden response format
    properties:
      code:
        example: 4003
        type: integer
      message:
        example: Access denied
        type: string
    type: object
  response.InternalServerErrorResponse:
    description: Internal Server Error response format
    properties:
      code:
        example: 5000
        type: integer
      message:
        example: Internal server error
        type: string
    type: object
...
  response.PaginatedResponse:
    description: Paginated response format
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          type: object
        type: array
      limit:
        example: 10
        type: integer
      message:
        example: success
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  response.StandardResponse:
    description: Standard API response format
    properties:
      code:
        description: Response code (0 for success, non-zero for errors)
        example: 0
        type: integer
      data:
        description: Response data (optional)
        type: object
      message:
        description: Response message
        example: success
        type: string
    type: object
  response.UnauthorizedResponse:
    description: Unauthorized response format
    properties:
      code:
        example: 4001
        type: integer
      message:
        example: Authentication required
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive service management platform with subscription-based
    billing, user management, and server administration. Features include OAuth2 authentication,
    traffic subscription management, multi-gateway payments, referral programs, and
    customer support system.
  title: Linke API
  version: "1.0"
paths:
...
  /admin/subscriptions/users:
    get:
      consumes:
      - application/json
      description: Get all user subscriptions with full details including user and
        subscription plan information (Admin only)
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by status
        enum:
        - active
        - inactive
        - cancelled
        - expired
        - trialing
        - past_due
        in: query
        name: status
        type: string
      - description: Filter by subscription plan ID
        in: query
        name: plan_id
        type: integer
      - description: Limit results
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Offset results
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserSubscriptionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] List all user subscriptions'
      tags:
      - Admin-Subscription-Management
    post:
      consumes:
      - application/json
      description: Create a new user subscription directly (Admin only)
      parameters:
      - description: User subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/admin.CreateUserSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Create user subscription'
      tags:
      - Admin-Subscription-Management
  /admin/subscriptions/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user subscription (Admin only)
      parameters:
      - description: User subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Delete user subscription'
      tags:
      - Admin-Subscription-Management
    get:
      consumes:
      - application/json
      description: Get a user subscription by ID with full details (Admin only)
      parameters:
      - description: User subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Get user subscription by ID'
      tags:
      - Admin-Subscription-Management
    patch:
      consumes:
      - application/json
      description: Partially update a user subscription with only provided fields
        (Admin only)
      parameters:
      - description: User subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription fields to update
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Partially update user subscription'
      tags:
      - Admin-Subscription-Management
    put:
      consumes:
      - application/json
      description: Update a user subscription (Admin only)
      parameters:
      - description: User subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Update user subscription'
      tags:
      - Admin-Subscription-Management
...
  /admin/tickets:
    get:
      consumes:
      - application/json
      description: Get all tickets in the system with advanced filtering options
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by assigned admin ID
        in: query
        name: assigned_to_id
        type: integer
      - description: Filter by status
        enum:
        - open
        - in_progress
        - pending
        - resolved
        - closed
        in: query
        name: status
        type: string
      - description: Filter by priority
        enum:
        - low
        - normal
        - high
        - urgent
        - critical
        in: query
        name: priority
        type: string
      - description: Filter by category
        enum:
        - general
        - technical
        - billing
        - account
        - feature
        - bug
        - subscription
        - payment
        in: query
        name: category
        type: string
      - description: Search in title, description, or ticket number
        in: query
        name: search
        type: string
      - default: 10
        description: Limit number of results
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tickets retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.TicketResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List all tickets (Admin)
      tags:
      - Admin-Ticket-Management
  /admin/tickets/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a ticket (admin only)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid ticket ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a ticket (Admin)
      tags:
      - Admin-Ticket-Management
    get:
      consumes:
      - application/json
      description: Get details of a specific ticket by ID with full admin access
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TicketResponse'
              type: object
        "400":
          description: Invalid ticket ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get a specific ticket (Admin)
      tags:
      - Admin-Ticket-Management
    put:
      consumes:
      - application/json
      description: Update ticket details (admin only)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ticket updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TicketResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a ticket (Admin)
      tags:
      - Admin-Ticket-Management
...
  /admin/tickets/number/{ticket_no}:
    get:
      consumes:
      - application/json
      description: Get details of a specific ticket by ticket number with full admin
        access
      parameters:
      - description: Ticket Number
        in: path
        name: ticket_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TicketResponse'
              type: object
        "400":
          description: Invalid ticket number
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get a ticket by ticket number (Admin)
      tags:
      - Admin-Ticket-Management
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"